Move utils to own package
