+++ engine/util/maps.go
@@ -1,4 +1,4 @@
-package engine
+package util
 
 // Keys returns the keys of the map m.
 // The keys will be an indeterminate order.
